backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
asyncio
retell-sdk
requests
openai
